requirements-airflow.txt
requests
python-dotenv
minio
kafka-python
elasticsearch
pandas
pyarrow
psycopg2-binary